+++ src/sandbox_test/scene_simple_interact.cpp
@@ -7,9 +7,9 @@ using namespace war;
 
 static Bit_Flag_Generator collision_bits = 1;
 
-static const unsigned scene_simple_coll_player = collision_bits.get();
-static const unsigned scene_simple_coll_geo = collision_bits.next();
-static const unsigned scene_simple_coll_npc = collision_bits.next();
+static const uint16_t scene_simple_coll_player = collision_bits.get();
+static const uint16_t scene_simple_coll_geo = collision_bits.next();
+static const uint16_t scene_simple_coll_npc = collision_bits.next();
 
 // ----------------------------------------------------------------------------
